docker_version/requirements.txt
flask
python-dotenv
PyAudio
sounddevice
pydub
google-genai
